backend/pyproject.toml
[project]
name = "datum"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "kreuzberg>=2.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1.0",
    "watchdog>=6.0.0",
    "pyyaml>=6.0.2",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.18",
    "sqlparse>=0.5.3",
    "tiktoken>=0.8.0",
    "mcp>=1.10.1,<2",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pip-audit>=2.8.0",
    "types-PyYAML>=6.0.12.20250915",
]

[project.scripts]
datum = "datum.cli.main:main"
datum-eval = "datum.cli.eval:main"
datum-insights = "datum.cli.insights:main"

[tool.setuptools.packages.find]
include = ["datum*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[[tool.mypy.overrides]]
module = ["frontmatter", "pymupdf4llm", "kreuzberg", "yaml", "pgvector.*"]
ignore_missing_imports = true
